refactor(client): extract DailyActiveRow from DailiesActiveListView
The list view's 181-line inline row callback becomes a named
DailyActiveRow component (fallow scores nested lambdas separately and
the clone detector can't see inline components), leaving the list as a
thin grid map. Stories pin the row with and without history.
